move functions to utils file
